changed structure of tests
